Added 'json' and 'js-string' as standard names, but didn't hardcode them into the Python version (see comments in code).
This is used to generate unit test methods in JavaScript.

What the existing javascript_verifier does is generate a JavaScript snippet and shell out
to the v8 shell each time to run it.

This instead generates a big HTML file with a bunch of tests inline, and
includes the JavaScript code and the test framework, and thus the tests can be
run in single page load of a browser.