For Python, implement the smart-indent behavior.  If "block" directive or a comment appears by a line by itself, omit all leading and trailing whitespace on the line.
This required changing the whitespace in 3 tests.  Since we haven't yet changed
the JavaScript and Java versions, I added ignore_all_whitespace=True to their
Expansion verifiers.

I don't think this should break anyone, but if it does, yell.  The benefit is
that the output will be a lot more readable -- useful for generating both
HTML/XML and code.