Implement the ability to format the current JSON node before template expansion.
This is useful for say splitting a list into columns -- that is part of the
presentation and thus should go in the template rather than the application.

For example, you would pass 'dirs': ['a', ... 'z'] and you may want a template
that looks like:

a b c
d e f
g h i
...

Example syntax:

{.repeated section dirs | columns}
  {.repeated section @}
    {@}
  {.end}
{.end}

It may make sense later to make "post-formatters", for formatting the string
after expansion.