Repurpose the {.foo?} syntax for attribute lookups.
Retained almost all backward compatibility.

Previously, {.if foo?} and {.foo?} were synonyms for looking up the predicate
'foo?'.

Now predicates should NOT have ? at the end of them.  The new recommendation for
predicates is:

{.if singular}
  one
{.or plural}
  multiple
{.end}

If you want to test an attribute named 'debug', you may use:

{.debug?}
  Rendered in {seconds} seconds
{.end}

rather than:

{.if test debug}
  Rendered in {seconds} seconds
{.end}


If there are multiple things to test in an if/or chain, just write it out:

{.if test debug}
  DEBUG
{.or test trace}
  TRACE
{.or}
  (none)
{.end}

Need to write a wiki page about this.