Python and JavaScript:
Be more lenient against embedded code.  The first character inside {...} should
be a nonspace, and there has to be at least one.

This is true for all our constructs:

{name}
{# comment}
{.section foo}, {.meta-left}, etc.

So it shouldn't break any constructs.

Previously we would interpret:

function() { return 1; }

as a template with the substitution " return 1; ".

Now it is treated as the literal string:

"function() { return 1; }"

In contrast,

function() {return 1;}

is still a template with the substitution "return 1;".